+++ examples/pointnet2_benchmark_segmentation.py
@@ -150,6 +150,7 @@ def main() -> None:
 
     print(f"Benchmarking model {args.model!r} on S3DIS (areas={args.areas})!")
     model, model_info = create_model(args.model, task="semantic-segmentation", pretrained=True, return_info=True)
+    assert model_info["transform"] is not None
     num_classes = int(model.num_classes)
     transform, inferer_transform, build_inferer = PROTOCOLS[args.model]
     inferer_transform = inferer_transform or model_info["transform"]