+++ examples/pointrcnn_benchmark_detection.py
@@ -110,6 +110,7 @@ def main() -> None:
 
     print(f"Benchmarking model {args.model!r} on KITTI!")
     model, model_info = create_model(args.model, task="detection", pretrained=True, return_info=True)
+    assert model_info["transform"] is not None
     assert isinstance(model, DetectionModel)
 
     dataset: Dataset = KITTI(