+++ src/torch_pointcloud/models/point_transformer_v3.py
@@ -831,9 +831,10 @@ class PointTransformerV3Encoder(nn.Module):
         x: OptTensor,
         pos_grid: Tensor,
         batch: Tensor,
-        return_intermediates: Literal[True],
         pos: OptTensor = None,
         condition: Optional[str] = None,
+        *,
+        return_intermediates: Literal[True],
     ) -> Tuple[Tensor, Tensor, Tensor, List[SerializedFeaturesDict]]: ...
 
     @overload
@@ -842,9 +843,9 @@ class PointTransformerV3Encoder(nn.Module):
         x: OptTensor,
         pos_grid: Tensor,
         batch: Tensor,
-        return_intermediates: Literal[False] = False,
         pos: OptTensor = None,
         condition: Optional[str] = None,
+        return_intermediates: Literal[False] = False,
     ) -> Tuple[Tensor, Tensor, Tensor]: ...
 
     def forward(
@@ -852,9 +853,9 @@ class PointTransformerV3Encoder(nn.Module):
         x: OptTensor,
         pos_grid: Tensor,
         batch: Tensor,
-        return_intermediates: bool = False,
         pos: OptTensor = None,
         condition: Optional[str] = None,
+        return_intermediates: bool = False,
     ) -> Any:
         x = x if x is not None else pos_grid.float()
 
@@ -1713,6 +1714,7 @@ def _ptv3_seg_transforms(relabel_labels: Optional[Sequence[int]] = None, estimat
 @register_model(
     "ptv3-base.scannet20.pointcept",
     task="semantic-segmentation",
+    input_keys=("x", "pos_grid", "batch"),
     weights=WeightsDict(
         url="hf://torch-pointcloud/ptv3-base.scannet20.pointcept/resolve/3a8e0a66d76ecfb57d4dd0e4ef43de18f1491b41/model.safetensors",
         dataset="scannet20",
@@ -1731,6 +1733,7 @@ def ptv3_base_scannet20(**hparams: Any) -> PointTransformerV3Segmentation:
 @register_model(
     "ptv3-base.scannet200.pointcept",
     task="semantic-segmentation",
+    input_keys=("x", "pos_grid", "batch"),
     weights=WeightsDict(
         url="hf://torch-pointcloud/ptv3-base.scannet200.pointcept/resolve/1f1232df7e827ff72dd7cd038763e75516494462/model.safetensors",
         dataset="scannet200",
@@ -1748,6 +1751,7 @@ def ptv3_base_scannet200(**hparams: Any) -> PointTransformerV3Segmentation:
 @register_model(
     "ptv3-base.s3dis-area5.pointcept",
     task="semantic-segmentation",
+    input_keys=("x", "pos_grid", "batch"),
     weights=WeightsDict(
         url="hf://torch-pointcloud/ptv3-base.s3dis-area5.pointcept/resolve/95d483ef0e186a94ffbf8302a2bba9b5ba710344/model.safetensors",
         dataset="s3dis-area5",